added /hi route
