--- src/main/java/EurekaServer.java
@@ -1,14 +1,0 @@
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.netflix.eureka.server.EnableEurekaServer;
-
-@SpringBootApplication
-@EnableEurekaServer
-public class EurekaServer {
-	
-	public static void main(String[] args) {
-		
-		SpringApplication.run(EurekaServer.class, args);
-	}
-	
-}
